+++ maskify.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  mode: 'mask',
-  fields: ['email', 'password'],
-  maskOptions: {
-    maskChar: '#',
-  },
-};